Get rid of the VirtualMatch file
