+++ opencog/query/PatternMatch.h
@@ -41,13 +41,6 @@ class PatternMatch
 	friend class SatisfactionLink;
 
 	protected:
-		// See PatternMatch.cc for comments
-		static void do_match(PatternMatchCallback *,
-		              const std::set<Handle>& vars,
-		              const HandleSeq& virtuals,
-		              const std::vector<HandleSeq>& components,
-		              const std::vector<std::set<Handle>>& comp_vars);
-
 		static bool recursive_virtual(PatternMatchCallback *cb,
 		            const std::vector<Handle>& virtuals,
 		            const std::vector<Handle>& negations,
